fix OpenAuthoringProvider circular dependency
