+++ packages/react-tinacms-github/src/github-error/OpenAuthoringErrorModal.tsx
@@ -18,7 +18,7 @@ limitations under the License.
 
 import { useEffect, useState } from 'react'
 import { useCMS } from 'tinacms'
-import { useOpenAuthoring } from '../open-authoring/OpenAuthoringProvider'
+import { useOpenAuthoring } from '../open-authoring/useOpenAuthoring'
 import { getModalProps } from './github-interpeter'
 import React from 'react'
 import {
