+++ packages/react-tinacms-github/src/open-authoring/OpenAuthoringContext.ts
@@ -1,0 +1,11 @@
+import React from 'react'
+
+export interface OpenAuthoringProps {
+  enterEditMode: () => void
+  exitEditMode: () => void
+  setError: (err: any) => void
+}
+
+export const OpenAuthoringContext = React.createContext<OpenAuthoringProps | null>(
+  null
+)
