+++ packages/react-tinacms-github/src/open-authoring/useOpenAuthoring.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import { OpenAuthoringContext } from './OpenAuthoringContext'
+
+export function useOpenAuthoring() {
+  const openAuthoringContext = React.useContext(OpenAuthoringContext)
+
+  if (!openAuthoringContext) {
+    throw new Error('useOpenAuthoring must be within an OpenAuthoringContext')
+  }
+
+  return openAuthoringContext
+}
